# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-commerce storefront API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List published products
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
...
  /products/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/NotFound" }
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      parameters:
        - name: cartId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, quantity]
              properties:
                cartId: { type: string }
                productId: { type: integer }
                quantity: { type: integer, minimum: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cartId, quantity]
              properties:
                cartId: { type: string }
                quantity: { type: integer, minimum: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
    delete:
      operationId: removeCartItem
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
        - name: cartId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
...
  /checkout:
    post:
      operationId: createCheckout
      tags: [orders]
      summary: Create Stripe checkout session for cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cartId, address, shippingRateId]
              properties:
                cartId: { type: string }
                email: { type: string, format: email }
                address: { $ref: "#/components/schemas/AddressInput" }
                shippingRateId: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [url, orderId]
                properties:
                  url: { type: string }
                  orderId: { type: integer }
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
    Product:
      type: object
      required: [id, slug, name, description, priceCents, currency, images, inventory, featured, published, createdAt]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        shortDescription: { type: ["string", "null"] }
        priceCents: { type: integer }
        compareAtCents: { type: ["integer", "null"] }
        currency: { type: string }
        images:
          type: array
          items: { type: string }
        inventory: { type: integer }
        weightOz: { type: ["number", "null"] }
        tags:
          type: array
          items: { type: string }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        featured: { type: boolean }
        published: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    Cart:
      type: object
      required: [id, items, subtotalCents, currency]
      properties:
        id: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        subtotalCents: { type: integer }
        currency: { type: string }
    CartItem:
      type: object
      required: [id, productId, quantity, product, lineTotalCents]
      properties:
        id: { type: integer }
        productId: { type: integer }
        quantity: { type: integer }
        product: { $ref: "#/components/schemas/Product" }
        lineTotalCents: { type: integer }
    AddressInput:
      type: object
      required: [name, street1, city, state, zip, country]
      properties:
        name: { type: string }
        street1: { type: string }
        street2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        zip: { type: string }
        country: { type: string }
        phone: { type: ["string", "null"] }
...
    Order:
      type: object
      required: [id, status, items, subtotalCents, shippingCents, taxCents, totalCents, currency, createdAt]
      properties:
        id: { type: integer }
        status: { type: string }
        email: { type: ["string", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        subtotalCents: { type: integer }
        shippingCents: { type: integer }
        taxCents: { type: integer }
        totalCents: { type: integer }
        currency: { type: string }
        shippingAddress: { $ref: "#/components/schemas/AddressInput" }
        trackingCode: { type: ["string", "null"] }
        labelUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, priceCents]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productImage: { type: ["string", "null"] }
        quantity: { type: integer }
        priceCents: { type: integer }
...